# Vectorize LowerCholeskyTransform