@vishwakftw 2019-08-10
I finally found use for the rocket emoticon.